+++ spec/auction_fun_core/entities/auction_spec.rb
@@ -4,7 +4,7 @@ require "spec_helper"
 
 RSpec.describe AuctionFunCore::Entities::Auction, type: :entity do
   describe "#initial_bid" do
-    subject(:auction) { Factory.structs[:auction] }
+    subject(:auction) { Factory.structs[:auction, :default_scheduled_standard] }
 
     it "expect return initial bid as money object" do
       expect(auction.initial_bid).to be_a_instance_of(Money)
@@ -12,7 +12,7 @@ RSpec.describe AuctionFunCore::Entities::Auction, type: :entity do
   end
 
   describe "#minimal_bid" do
-    subject(:auction) { Factory.structs[:auction] }
+    subject(:auction) { Factory.structs[:auction, :default_scheduled_standard] }
 
     it "expect return minimal bid as money object" do
       expect(auction.minimal_bid).to be_a_instance_of(Money)
@@ -21,7 +21,7 @@ RSpec.describe AuctionFunCore::Entities::Auction, type: :entity do
 
   describe "#winner?" do
     context "when there is an associated FK" do
-      subject(:auction) { Factory.structs[:auction, :with_winner, winner_id: 1] }
+      subject(:auction) { Factory.structs[:auction, :default_finished_standard, :with_winner, winner_id: 1] }
 
       it "expects to return true when it has a winning user associated." do
         expect(auction.winner?).to be_truthy
@@ -29,7 +29,7 @@ RSpec.describe AuctionFunCore::Entities::Auction, type: :entity do
     end
 
     context "when there is no associated FK" do
-      subject(:auction) { Factory.structs[:auction] }
+      subject(:auction) { Factory.structs[:auction, :default_scheduled_standard] }
 
       it "expect return a user object" do
         expect(auction.winner?).to be_falsey